+++ ssr/registerRouterSSR/index.d.ts
@@ -2,4 +2,4 @@ import Router from 'easyroute-core'
 
 declare function registerRouter(router: Router): void
 
-export { registerRouter }
+export default registerRouter